requirements/base.txt
Django==2.0.2
djangorestframework==3.7.7
django-taggit==0.22.2
